requirements.txt
numpy
pandas
pyBigWig
tqdm
hurst
numba
MFDFA
scikit-learn
matplotlib
pyqt5
statsmodels
umap-learn
pyranges